feat: major changes
This project was previously kept in [this
repo](https://github.com/mikededo/stack), and it has been moved into
it's own repository.
Commits have been kept through `git filter-repo` but there were some
modifications missing to be made, mostly configurations files.
The project will be published in vercel, same as it was previously.